remember.el causes a nasty bug when capturing in OrgMode, leaving stray
buffers behind, with broken keybindings. We don't need remember anyway -
we'll just set the initial buffer to our notes file instead.

remember also has a major annoyance in that it is a minor mode and it
overwrites the very useful OrgMode C-c C-c binding. Getting rid of
remember is better in multiple ways.