Emacs 25 has a very stupid 'feature' of modifying your init.el file.
package.el in Emacs 25 adds (package-initialize) with some commentary as
to why, when it is not found anywhere in the file. This is a poor design
choice and should be fixed above all other bugs. Temporarily, we will
apply a hack to disable this feature, which may break in the future as
it uses an internal variable, hence being a hack. We do not need init.el
modified ever, and we do not need it in init.el anyway, as our main
configuration is config.el.